requirements.txt
streamlit
pandas
plotly
Pillow
numpy
streamlit-aggrid